requirements.txt
numpy==2.1
Pillow==10.4.0
